Changes to Delta calculation
Different conventions exist as to how delta is calculated. WVASE calculates delta between 0 and 180°; and value greater than 180° is 'reflected' around 180°. I think this is because ellipsometers without a compensator are unable to distinguish between the two cases.

Further investigation is required into specific ellipsometer types and the values of delta and psi they produce.

Currently, Refellips is behaving like WVASE, which has always worked for us in the past.