Allow wavelength to be set for RI
Allows for the user to specify a wavelength for the RI to use - this is useful for cases where you are simply trying to plot a refractive index profile without a model attached (i.e., just a structure). It will also be useful for troubleshooting.

If a model is linked to the refractive index, that will take priority over any wavelength provided through set_wav